requirements.txt
pyrtlsdr==0.3.0
numpy==1.24.3
scipy==1.10.1
pyyaml==6.0.1